[nomerge] Try un-reverting except on TextInput code -- bug returns.
If we can, we'd like to get the benefit of the original change at
least on ordinary Text components.  The issue we're fixing seems like
it can't trigger there, or at least is much less likely to than on
TextInput.

So, try un-reverting all the changes to files that aren't specifically
about TextInput.

Unfortunately, this causes the issue to reproduce again.