[nomerge] Try re-reverting build-from-shadow instead -- now works!
Undo the last commit's experiment, and re-revert instead the other
spot where the original commit caused its new code to get called:
inside `buildSpannedFromShadowNode`.

This causes the issue to be fixed again!  So, this hunk is the
difference between TextInputs getting (painfully) slow and not.

Unfortunately, it also looks to me like it's probably the key entry
point to the whole change -- so keeping this knocked out has much the
same effect in practice as just reverting the entire original commit.