wip Add a $ReadOnly to enable heterogeneous SectionLists
TODO finish commit message; use "explicit section type" test case,
  which I thought of when trying to simplify this explanation

This type is naturally read-only, because it's the type of some data
that's passed via props to SectionList and its friends.

The effect of saying so is that it becomes possible to have a SectionT
which is a union type -- so e.g. some sections have one type of item
and certain extra properties, while other sections have a different
type of item and different extra properties -- while plumbing all the
types soundly through, without resorting to $FlowFixMe.  A small new
flowtest case demonstrates this.

In the flowtest example, the two branches of the union are subtypes of
SectionBase<string> and SectionBase<number> respectively.  With this
change, their union is therefore a subtype of SectionBase<string | number>
and everything works.

Each branch
of the union is a subtype of some SectionBase<T> for a different T.
Without $ReadOnly, the

Each branch
of the union is a subtype of SectionBase<SectionItemT> with a different
type argument SectionItemT; but without the $ReadOnly,